// src/store/enums/StoreEnums.ts
enum Actions {
  // action types
  GET_MOVIE = "getMovieDetail",
  GET_MOVIE_CREDITS = "getMovieCredits",
  GET_MOVIES = "getww",
  GET_GENRES = "getGenre",
  SEND_RATE = "sendRate",
}

enum Mutations {
  // mutation types
  SET_MOVIES = "setMovies",
  SET_MOVIE = "setww",
  SET_MOVIE_CREDITS = "setMovieCredits",
  SET_GENRES = "setGenre",
  SET_ERROR = "setError",
}

export { Actions, Mutations };